+++ src/Resources/ResourceManager/SdkClient/ResourceManagerSdkClient.cs
@@ -1104,14 +1104,37 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
 
         public virtual PSResource GetById(string resourceId, string apiVersion)
         {
-            PSResource result = null;
+            var providers = new List<Provider>();
             var resourceIdentifier = new ResourceIdentifier(resourceId);
-            var providers = ResourceManagementClient.Providers.List();
+            var providerNamespace = ResourceIdentifier.GetProviderFromResourceType(resourceIdentifier.ResourceType);
+            if (!string.IsNullOrEmpty(providerNamespace))
+            {
+                var result = ResourceManagementClient.Providers.Get(providerNamespace);
+                if (result != null)
+                {
+                    providers.Add(result);
+                }
+            }
+
+            if (!providers.Any())
+            {
+                var result = ResourceManagementClient.Providers.List();
+                if (result != null)
+                {
+                    result.ForEach(p => providers.Add(p));
+                    while (!string.IsNullOrEmpty(result.NextPageLink))
+                    {
+                        result = ResourceManagementClient.Providers.ListNext(result.NextPageLink);
+                        result.ForEach(p => providers.Add(p));
+                    }
+                }
+            }
+
             foreach (var provider in providers)
             {
                 var resourceType = provider.ResourceTypes
-                                            .Where(t => string.Equals(string.Format("{0}/{1}", provider.NamespaceProperty, t.ResourceType), resourceIdentifier.ResourceType, StringComparison.OrdinalIgnoreCase))
-                                            .FirstOrDefault();
+                                           .Where(t => string.Equals(string.Format("{0}/{1}", provider.NamespaceProperty, t.ResourceType), resourceIdentifier.ResourceType, StringComparison.OrdinalIgnoreCase))
+                                           .FirstOrDefault();
                 if (resourceType != null)
                 {
                     apiVersion = resourceType.ApiVersions.Contains(apiVersion) ? apiVersion : resourceType.ApiVersions.FirstOrDefault();
@@ -1122,7 +1145,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.SdkClient
                 }
             }
 
-            return result;
+            return null;
         }
     }
 }