+++ src/Accounts/Accounts/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated Add-AzEnvironment and Set-AzEnvironment to accept parameter AzureAnalysisServicesEndpointResourceId
+* Bug fix for azure automation update management dynamic group
 
 ## Version 1.4.0
 * Add 'Register-AzModule' command to support AutoRest generated cmdlets